moves components into src
